refactor: rmv unsed route, moce chat page to index and separated styles chat
