--- src/routes/_public/about.tsx
@@ -1,9 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/_public/about")({
-  component: About,
-});
-
-function About() {
-  return <div className="p-2">Hello from About!</div>;
-}
